+++ app/models/his/his_haos.js
@@ -632,7 +632,7 @@ class HisHaosModel {
             sqlCommand.where(columnName, searchValue);
         }
         if (columnName == 'i.dchdate') {
-            sqlCommand.whereRaw('LENGTH(i.rfrilct)=5');
+            sqlCommand.whereRaw('LENGTH(i.rfrilct) IN (5,9)');
         }
         return sqlCommand
             .select(db.raw(`
@@ -710,7 +710,7 @@ class HisHaosModel {
         columnName = columnName === 'visitNo' ? 'q.vn' : columnName;
         columnName = columnName === 'dateadmit' ? 'i.regdate' : columnName;
         columnName = columnName === 'datedisc' ? 'i.dchdate' : columnName;
-        let validRefer = columnName === 'datedisc' ? ' AND LENGTH(i.rfrilct)=5 ' : '';
+        let validRefer = columnName === 'datedisc' ? ' AND LENGTH(i.rfrilct) IN (5,9) ' : '';
         const sql = `
             SELECT
                 (select hospitalcode from opdconfig) as HOSPCODE,
@@ -1265,7 +1265,7 @@ class HisHaosModel {
             .select(db.raw(`concat(ovst.vstdate, ' ',ovst.vsttime) as DATETIME_IN, '1' as REFER_RESULT`))
             .select(db.raw(`concat(ovst.vstdate, ' ',ovst.vsttime) as D_UPDATE`))
             .where(db.raw(`(referin.refer_date='${visitDate}' or referin.date_in='${visitDate}')`))
-            .where(db.raw('length(referin.refer_hospcode)=5'))
+            .where(db.raw('length(referin.refer_hospcode) IN (5,9)'))
             .whereNotNull('referin.vn')
             .whereNotNull('patient.hn')
             .limit(maxLimit);