+++ app/task/moph-erp.js
@@ -38,6 +38,7 @@ const sendBedOccupancy = async (dateProcess = null) => {
         ]);
         dateOpd = moment(dateOpd).add(1, 'day').format('YYYY-MM-DD');
     } while (dateOpd <= currDate);
+    console.log('-'.repeat(70));
     return { clinicResult, wardResult, opdResult };
 };
 exports.sendBedOccupancy = sendBedOccupancy;
@@ -52,10 +53,12 @@ const sendBedOccupancyByWard = async (date) => {
             const result = await (0, moph_refer_1.sendingToMoph)('/save-occupancy-rate-by-ward', rows);
             console.log(moment().format('HH:mm:ss'), 'send Occ Rate by ward', date, result.status || '', result.message || '', rows.length, 'rows');
         }
+        console.log('-'.repeat(70));
         return rows;
     }
     catch (error) {
         console.error(moment().format('HH:mm:ss'), 'sendBedOccupancy error by ward', date, error.message);
+        console.log('-'.repeat(70));
         return false;
     }
 };
@@ -138,15 +141,18 @@ const sendWardName = async () => {
             });
             const result = await (0, moph_refer_1.sendingToMoph)('/save-ward', rows);
             console.log(moment().format('HH:mm:ss'), 'sendWardName', result.status || '', result.message || '', rows.length, 'rows');
+            console.log('-'.repeat(70));
             return result;
         }
         else {
             console.log(moment().format('HH:mm:ss'), 'sendWardName', 'No ward data');
+            console.log('-'.repeat(70));
             return { statusCode: 200, message: 'No ward data' };
         }
     }
     catch (error) {
         console.log(moment().format('HH:mm:ss'), 'getWard error', error.message);
+        console.log('-'.repeat(70));
         return [];
     }
 };
@@ -186,10 +192,12 @@ const sendBedNo = async () => {
             times++;
         } while (startRow < countBed && countBed != 0);
         console.log(moment().format('HH:mm:ss'), `sendBedNo ${countBed} rows (${times} times)`, error);
+        console.log('-'.repeat(70));
         return { statusCode: 200, sentResult };
     }
     catch (error) {
         console.log(moment().format('HH:mm:ss'), 'getBedNo error', error.message);
+        console.log('-'.repeat(70));
         return { statusCode: error.status || 500, message: error.message || error };
     }
 };
@@ -220,10 +228,12 @@ const updateAlive = async () => {
         else {
             console.error(moment().format('HH:mm:ss'), '❌ Sent API Alive status result', result.status || '', result.statusCode || '', result?.message || '');
         }
+        console.log('-'.repeat(70));
         return result;
     }
     catch (error) {
         console.error(moment().format('HH:mm:ss'), '❌ Sent API Alive status error:', error?.status || error?.statusCode || '', error?.message || error || '');
+        console.log('-'.repeat(70));
         return [];
     }
 };
@@ -271,10 +281,12 @@ const erpAdminRequest = async () => {
         else {
             console.log(moment().format('HH:mm:ss'), 'No admin request', result.status || result?.statusCode || '', result?.data?.message || result?.message || '');
         }
+        console.log('-'.repeat(70));
         return result;
     }
     catch (error) {
         console.log(moment().format('HH:mm:ss'), 'Admin Request error', error.message);
+        console.log('-'.repeat(70));
         return [];
     }
 };