+++ src/models/his/his_hosxpv3.ts
@@ -827,7 +827,7 @@ export class HisHosxpv3Model {
       sqlCommand.where(columnName, searchValue);
     }
     if (columnName == 'i.dchdate') {
-      sqlCommand.whereRaw('LENGTH(i.rfrilct)=5'); // get only referin
+      sqlCommand.whereRaw('LENGTH(i.rfrilct) IN (5,9)'); // get only referin
     }
     return sqlCommand
       .select(db.raw(`
@@ -1410,7 +1410,7 @@ export class HisHosxpv3Model {
       .select(db.raw(`concat(ovst.vstdate, ' ',ovst.vsttime) as DATETIME_IN, '1' as REFER_RESULT`))
       .select(db.raw(`concat(ovst.vstdate, ' ',ovst.vsttime) as D_UPDATE`))
       .where(db.raw(`(referin.refer_date=? or referin.date_in=?)`, [visitDate, visitDate]))
-      .where(db.raw('length(referin.refer_hospcode)=5'))
+      .where(db.raw('length(referin.refer_hospcode) IN (5,9)'))
       .whereNotNull('referin.vn')
       .whereNotNull('patient.hn')
       .limit(maxLimit);